refactor(runtime): extract archive helpers; infer bash need for startSurface processes
- downloadFile/extractZip move from androidInstaller.ts to a dependency-light
  archiveUtils.ts, so windowsBash (and installer.ts's MINGIT_VERSION import)
  no longer transitively load the Android installer on every consumer.
- inferRuntimeNeeds counts startSurface process descriptors toward the
  windowsBash need — they run through the same launcher and config-level
  shell default as runShell.background.
